pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claw-cli-demo"
version = "0.1.0"
description = "Python CLI demo with skills and grep-style memory"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.6.0",
  "PyYAML>=6.0.0",
  "python-dotenv>=1.0.0",
  "openai>=1.0.0",
  "requests>=2.31.0",
]

[project.scripts]
claw = "claw_demo.main:run"

[tool.setuptools.packages.find]
include = ["claw_demo*"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
